openchrome doesn't build if you have an xserver with DRI compiled in, but you disable dri on in the driver build and don't have the via_drm.h it wants in dri mode. Work around this by changing the #define used to an openchrome only one, instead of overloading the XF86DRI identifier.
Also disable the DRI build unconditionally.

This is needed here since we don't provide the via DRM module, and i'm
moving libdrm over to using the kernel headers (instead of libdrms own
private copy of same... This is why kernel modules should be developed
in kernel). We won't provide a via drm driver until it is re-written,
since it is full of linuxisms (like futex).

ok matthieu@, discussed with a few. tested by grange@ to prove it was a
no-op functionality wise.